# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Compliance Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /fire-safety:
    get:
      operationId: listFireSafetyChecks
      tags: [fire-safety]
      summary: List fire safety logbook entries
      parameters:
        - in: query
          name: checkType
          schema:
            $ref: "#/components/schemas/FireCheckType"
        - in: query
          name: siteId
          schema:
            type: integer
      responses:
        "200":
          description: List of fire safety checks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FireSafetyCheck"
    post:
      operationId: createFireSafetyCheck
      tags: [fire-safety]
      summary: Record a fire safety check
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFireSafetyCheckRequest"
      responses:
        "201":
          description: Created check
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FireSafetyCheck"
  /fire-safety/status:
    get:
      operationId: getFireSafetyStatus
      tags: [fire-safety]
      summary: Per-check-type due status (last done, next due, overdue flags)
      parameters:
        - in: query
          name: siteId
          schema:
            type: integer
      responses:
        "200":
          description: Status per check type
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FireSafetyStatus"
  /fire-safety/{id}:
    put:
      operationId: updateFireSafetyCheck
      tags: [fire-safety]
      summary: Update a fire safety check entry
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFireSafetyCheckRequest"
      responses:
        "200":
          description: Updated check
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FireSafetyCheck"
    delete:
      operationId: deleteFireSafetyCheck
      tags: [fire-safety]
      summary: Delete a fire safety check entry
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /fire-safety/config:
    get:
      operationId: getFireSafetyConfig
      tags: [fire-safety]
      summary: Get Fire safety configuration
      responses:
        "200":
          description: Fire safety configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FireSafetyConfig"
    put:
      operationId: updateFireSafetyConfig
      tags: [fire-safety]
      summary: Update Fire safety configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FireSafetyConfig"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FireSafetyConfig"
...
components:
  schemas:
...
    FireCheckType:
      type: string
      enum: [alarm, emergency_lights, extinguishers, fire_doors, fire_drill]
    FireSafetyCheck:
      type: object
      properties:
        id:
          type: integer
        clientId:
          type: integer
        siteId:
          type: integer
          nullable: true
        checkType:
          $ref: "#/components/schemas/FireCheckType"
        checkDate:
          type: string
        result:
          type: string
          enum: [pass, fail]
        location:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        performedBy:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - clientId
        - checkType
        - checkDate
        - result
        - createdAt
        - updatedAt
    CreateFireSafetyCheckRequest:
      type: object
      properties:
        checkType:
          $ref: "#/components/schemas/FireCheckType"
        checkDate:
          type: string
        result:
          type: string
          enum: [pass, fail]
        siteId:
          type: integer
          nullable: true
        location:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        performedBy:
          type: string
          nullable: true
      required:
        - checkType
        - checkDate
        - result
    UpdateFireSafetyCheckRequest:
      type: object
      properties:
        checkDate:
          type: string
        result:
          type: string
          enum: [pass, fail]
        siteId:
          type: integer
          nullable: true
        location:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        performedBy:
          type: string
          nullable: true
    FireSafetyStatus:
      type: object
      properties:
        checkType:
          $ref: "#/components/schemas/FireCheckType"
        frequencyDays:
          type: integer
        lastDate:
          type: string
          nullable: true
        lastResult:
          type: string
          nullable: true
        dueDate:
          type: string
          nullable: true
        status:
          type: string
          enum: [ok, due_soon, overdue, never]
      required:
        - checkType
        - frequencyDays
        - status
    FireSafetyConfig:
      type: object
      properties:
        fire_alarm_zones:
          type: string
        fire_extinguisher_points:
          type: string
        fire_show_drill:
          type: string
        fire_default_performer:
          type: string
        fire_escape_routes:
          type: string
...